Refactoring: Moved Task files to the folder /tasks. Created a file math.h for calculations such as altitude
