+++ AvionicsSoftware-AtollicProject/Inc/tasks/cli.h
@@ -66,7 +66,7 @@ typedef struct{
 // Returns:
 //   VOID
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
-void vTask_xtract(void *pvParameters);
+void vTask_CLI(void *pvParameters);
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // Description:
