+++ AvionicsSoftware-AtollicProject/Src/tasks/pressure_sensor_bmp280.c
@@ -16,7 +16,7 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
-#include <pressure_sensor_bmp280.h>
+#include <tasks/pressure_sensor_bmp280.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
