+++ AvionicsSoftware-AtollicProject/Src/tasks/pressure_sensor_bmp3.c
@@ -16,7 +16,7 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
-#include <pressure_sensor_bmp3.h>
+#include <tasks/pressure_sensor_bmp3.h>
 #include <stdlib.h>
 
 #include "bmp3.h"
