+++ AvionicsSoftware-AtollicProject/Src/tasks/sensorAG.c
@@ -18,7 +18,7 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
-#include "sensorAG.h"
+#include "tasks/sensorAG.h"
 #include "bmi088.h"
 #include "SPI.h"
 #include "hardwareDefs.h"
