+++ AvionicsSoftware-AtollicProject/Src/tasks/startupTask.c
@@ -15,7 +15,7 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
-#include "startupTask.h"
+#include "tasks/startupTask.h"
 
 #include "cmsis_os.h"
 #include "configuration.h"
